Publish live site if it does not exist
Publish live site with existing archive unconditionally if live site
does not exist even if there is no new top story for TopHN since the
last update to archive database.

This is done to accomodate a scenario where the live site is deleted
(for example to re-publish it with new HTML changes) and we need to
publish a new live site immediately although there is no update to the
archive database.

Apart from this functional change, there are minor code changes related
to code formatting and logging.