Add HTML5 DOCTYPE declaration
Additionally, fix an existing bug in the stylesheet that got exposed
after adding the HTML5 DOCTYPE declaration.

The stylesheet uses flexbox to vertically center the top story in the
home page. However, after adding HTML5 DOCTYPE declaration, the top
story is no longer vertically centered but instead moved to the top of
the page. This happens because the top story is vertically centered
within its parent, i.e. HTML body, whose height is only as tall as its
content after adding the DOCTYPE declaration.

Fix this issue by setting the height of HTML body to 100%.